backend/requirements.txt
flask
torch
transformers
Pillow
opencv-python
tqdm
scikit-image
requests
huggingface-hub
flask-cors